refactor(modal): split the 3.6k-line _app.py into cohesive modules
`integration/modal/_app.py` had accumulated the entire Modal integration —
app, executor, builder, runner, reactive bootstrap, scheduler tick, worker
selection, executor metadata, profile config and volume discovery — in 3606
lines. Split it along the seams the code already had, leaving `_app.py` with
`StardagApp` alone:

  _logging     15   container logging setup
  _volumes     73   modalvol:// target roots -> modal.Volume
  _selector    79   worker routing (WorkerSelector, WorkerSelectorByName)
  _settings    99   FunctionSettings + the deploy-time merge
  _profile    112   stardag profile -> container env vars / Secret
  _protocols  120   BuildFunction / RunFunction contracts
  _builder    226   Builder, PrefectBuilder (resident orchestrator)
  _metadata   244   Modal coordinates + the env-var channel to workers
  _bootstrap  296   reactive bootstrap: discover, persist, arm, first tick
  _tick       433   tick body, tick config, watchdog sweep
  _runner     485   Runner + worker-side lifecycle self-reporting
  _executor   541   ModalTaskExecutor (spawn / re-attach / cancel)
  _app       1239   StardagApp

Behaviour is unchanged. Beyond the moves:

- `finalize()` drops from ~450 lines to ~190. The 135-line `_modal_tick`
  body moved to `_tick._run_tick`, and its eight deploy-time captures are
  bundled in a frozen `_TickDeployment` so the registered wrapper is a
  two-line delegation. Secret assembly, volume auto-mounting, the API-key
  check and per-worker timeouts became named helpers, and one local
  `register()` replaces five repetitions of the prepare-settings/name/
  serialized dance.
- The module map and usage example now live in the package `__init__`, the
  front door, instead of in a private module's header.
- `ModalTaskExecutor` is exported from `stardag.integration.modal`. It is
  the documented first-class executor and both the docs and a docstring
  example already told users to import it from there, but it was never in
  `__all__`.
- Dropped a redundant `import asyncio` inside `PrefectBuilder.build`.

Verification, since a refactor this size deserves more than a green suite:

- An AST diff of old vs new shows 136 of 144 symbols byte-identical modulo
  docstrings. The 8 that differ are `finalize` plus five annotation-only
  changes and the redundant-import removal.
- A probe capturing every `modal_app.function(**kwargs)` call that
  `finalize()` makes produces byte-identical output on both checkouts, for
  a minimal app and a fully-configured one: same registration order, names,
  secrets, volumes, timeouts and schedule, same `FinalizeResult`.
- New test: every registered function is `serialized=True`, i.e. Modal
  cloudpickles the closure at deploy. `TestDeployedFunctionsAreSerializable`
  round-trips all five through `modal._serialization` and drives the
  restored tick to prove its captures survive. Nothing covered this before,
  and it is what bundling the captures put at risk.

Test-side changes are import-path updates. Patch targets moved with the
code that reads them, and the hermetic-metadata fixture now patches each
helper in every module that calls it, not just where it is defined.
